+++ appconfigservice/api/controllers/app.py
@@ -15,6 +15,7 @@
 import logging
 import re
 from time import gmtime
+import yaml
 
 import auth_middleware
 import controllers.config as cfg
@@ -32,6 +33,26 @@ logging.basicConfig(level=logging.INFO, datefmt='%Y-%m-%dT%H:%M:%S',
                     format='%(asctime)-15s.%(msecs)03dZ %(levelname)-7s [%(threadName)-10s] : %(name)s - %(message)s')
 __logger = logging.getLogger("app_config_building_block")
 app = flask.Flask(__name__)
+
+
+def configs_ok_search():
+    """
+    Function for healthcheck public endpoint
+    Return : {'success':'true'}
+    """
+    msg = {"success": "true"}
+    return msg
+
+
+def configs_version_search():
+    """
+    Method to return AppConfig BB version number
+    Reads yaml file and returns version number
+    Return : plain text - version number
+    """
+    appconfig_yaml = open('appconfig.yaml')
+    parsed_appconfig_yaml = yaml.load(appconfig_yaml, Loader=yaml.FullLoader)
+    return parsed_appconfig_yaml['info']['version']
 
 
 def configs_search(mobileAppVersion=None):